+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Introduce a custom test classpath configuration for custom `TestableAware`-like tasks
 - Add `AddDefaultIntelliJPlatformDependencies` property to control if default dependencies (`com.intellij` and `intellij.rider`) are added automatically with IntelliJ Platform
 - Use the exact IntelliJ Platform version to resolve platform dependencies when targeting IntelliJ Platform from the nightly channel.
+- Introduce the `intellijPlatformTestRuntimeClasspath` configuration
 
 ### Changed